requirements.txt
python-telegram-bot
pyyaml
flask
cachetools>=5.3.0 